requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.3
Django==5.2.6
django-filter==25.1
et_xmlfile==2.0.0
lxml==6.0.1
mysqlclient==2.2.7
openpyxl==3.1.5
pillow==11.3.0
python-docx==1.2.0
reportlab==4.4.3
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
